{"schema":"decodex/commit/1","summary":"split production module boundaries","authority":"manual"}
